M2-4953: Add conftest for unit tests
