+++ src/apps/authentication/tests/unit/test_auth_deps.py
@@ -14,7 +14,6 @@ from apps.authentication.domain.token import InternalToken, JWTClaim
 from apps.authentication.domain.token.internal import TokenPurpose
 from apps.authentication.errors import AuthenticationError, InvalidCredentials
 from apps.authentication.services import AuthenticationService
-from apps.authentication.services.core import TokensService
 from apps.users.cruds.user import UsersCRUD
 from apps.users.db.schemas import UserSchema
 from apps.users.domain import User
@@ -23,47 +22,6 @@ from config import settings
 
 TEST_PASSWORD = "Test1234!"
 RJTI = str(uuid.uuid4())
-
-
-@pytest.fixture
-def token_blacklist_service(session: AsyncSession) -> TokensService:
-    return TokensService(session)
-
-
-@pytest.fixture
-def auth_service(session: AsyncSession) -> AuthenticationService:
-    return AuthenticationService(session)
-
-
-@pytest.fixture
-def access_token(user: User, auth_service: AuthenticationService) -> str:
-    data = {JWTClaim.sub: str(user.id), JWTClaim.rjti: RJTI}
-    return auth_service.create_access_token(data)
-
-
-@pytest.fixture
-def refresh_token(user: User, auth_service: AuthenticationService) -> str:
-    data = {JWTClaim.sub: str(user.id), JWTClaim.rjti: RJTI}
-    return auth_service.create_refresh_token(data)
-
-
-@pytest.fixture
-def access_token_internal(access_token: str, auth_service: AuthenticationService) -> InternalToken:
-    payload = auth_service.extract_token_payload(access_token, settings.authentication.access_token.secret_key)
-    return InternalToken(payload=payload)
-
-
-@pytest.fixture
-def refresh_token_internal(refresh_token: str, auth_service: AuthenticationService) -> InternalToken:
-    payload = auth_service.extract_token_payload(refresh_token, settings.authentication.refresh_token.secret_key)
-    return InternalToken(payload=payload)
-
-
-@pytest.fixture
-def mock_ws(mocker: MockerFixture):
-    mock = mocker.MagicMock(spec="fastapi.websockets.WebSocket")
-    mock.headers = {}
-    return mock
 
 
 async def test_get_current_user_for_ws__no_protocol_header(session: AsyncSession, mock_ws: WebSocket):
