+++ src/apps/authentication/tests/unit/test_auth_service.py
@@ -6,7 +6,7 @@ from pytest_mock import MockerFixture
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from apps.authentication.domain.login import UserLoginRequest
-from apps.authentication.domain.token import InternalToken, JWTClaim
+from apps.authentication.domain.token import InternalToken
 from apps.authentication.domain.token.internal import TokenPurpose
 from apps.authentication.errors import BadCredentials, InvalidCredentials
 from apps.authentication.services import AuthenticationService
@@ -14,44 +14,9 @@ from apps.authentication.services.core import TokensService
 from apps.users.cruds.user import UsersCRUD
 from apps.users.domain import User
 from apps.users.errors import UserIsDeletedError, UserNotFound
-from config import settings
 
 TEST_PASSWORD = "Test1234!"
 RJTI = str(uuid.uuid4())
-
-
-@pytest.fixture
-def token_blacklist_service(session: AsyncSession) -> TokensService:
-    return TokensService(session)
-
-
-@pytest.fixture
-def auth_service(session: AsyncSession) -> AuthenticationService:
-    return AuthenticationService(session)
-
-
-@pytest.fixture
-def access_token(user: User, auth_service: AuthenticationService) -> str:
-    data = {JWTClaim.sub: str(user.id), JWTClaim.rjti: RJTI}
-    return auth_service.create_access_token(data)
-
-
-@pytest.fixture
-def refresh_token(user: User, auth_service: AuthenticationService) -> str:
-    data = {JWTClaim.sub: str(user.id), JWTClaim.rjti: RJTI}
-    return auth_service.create_refresh_token(data)
-
-
-@pytest.fixture
-def access_token_internal(access_token: str, auth_service: AuthenticationService) -> InternalToken:
-    payload = auth_service.extract_token_payload(access_token, settings.authentication.access_token.secret_key)
-    return InternalToken(payload=payload)
-
-
-@pytest.fixture
-def refresh_token_internal(refresh_token: str, auth_service: AuthenticationService) -> InternalToken:
-    payload = auth_service.extract_token_payload(refresh_token, settings.authentication.refresh_token.secret_key)
-    return InternalToken(payload=payload)
 
 
 class TestAuthService:
